requirements.txt
nltk==3.2.5
pandas==0.22.0
Django==2.0.3
numpy==1.14.0
tensorflow==1.8.0
h5py==2.7.1
djangorestframework==3.8.2
tflearn==0.3.2
